Input Bertramka info
This commit should serve as the template for all parks data, however i believe only one park is necesary for the ICA1. The hashmap parks_info will eventually contain sub-hashmaps for every park on the website. All the info is structured in a way that the user can ask numerous questions where available such as.
Can I take a (tram, bus, metro) there?
What (tram, bus, metro) can I take?
What type of parking is there?
Where is the parking?
What are the opening hourse in (month)?

etc.

Commas in this map have been ommited and replaced by newlines. I think this looks much better but it does not have to stay that way.

--MARTIN november 1 10:30 - 11:20 am